+++ spec/dummy_app/lib/data_shifts/00000002_rake_spec_load_fail.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+raise "RAKE_SPEC_LOAD_BOOM"